fix: remove offending SEO code #99
While this may appear as a breaking change, this is actually the fix to
an issue where the SEO code was causing misreporting of SEO status. The
'fix' is to remove the offending SEO Code.

There is no backwards compatibility issue.

Changes to be committed:
modified:   src/auro-flight-main.js
modified:   test/auro-flight-main.test.js